# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Minecraft Nexus Panel (MCNP) API
  version: 0.1.0-draft
  summary: Panel API for Web, Desktop, and Mobile clients
  description: |
    Draft v1 contract. Browser clients use a secure session cookie plus CSRF
    token for writes. Native clients use bearer access tokens and rotating
    refresh tokens.
servers:
  - url: /api/v1
    description: Current Panel
...
security:
  - cookieAuth: [ ]
  - bearerAuth: [ ]
paths:
...
  /cores:
    get:
      tags: [ Cores ]
      operationId: listCores
      summary: List visible Core nodes
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CoreStatus'
      responses:
        '200':
          description: Cores visible to the current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorePage'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [ Cores ]
      operationId: createCore
      summary: Register a Core node
      security:
        - cookieAuth: [ ]
          csrfToken: [ ]
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoreCreate'
      responses:
        '201':
          description: Core registered; secret is never returned
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Core'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/CoreFailure'
  /cores/{coreId}:
    parameters:
      - $ref: '#/components/parameters/CoreId'
    get:
      tags: [ Cores ]
      operationId: getCore
      summary: Get a Core node
      responses:
        '200':
          description: Core details
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Core'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [ Cores ]
      operationId: updateCore
      summary: Update a Core registration
      security:
        - cookieAuth: [ ]
          csrfToken: [ ]
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/CoreUpdate'
      responses:
        '200':
          description: Updated Core
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Core'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/RevisionMismatch'
    delete:
      tags: [ Cores ]
      operationId: deleteCore
      summary: Remove a Core registration
      security:
        - cookieAuth: [ ]
          csrfToken: [ ]
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Core registration removed; remote data is untouched
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/RevisionMismatch'
  /cores/{coreId}/actions/test:
    post:
      tags: [ Cores ]
      operationId: testCoreConnection
      summary: Test Core connectivity and credentials
      security:
        - cookieAuth: [ ]
          csrfToken: [ ]
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/CoreId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                type: object
                required: [ success, latencyMs ]
                properties:
                  success:
                    type: boolean
                  latencyMs:
                    type: integer
                    minimum: 0
                  protocolVersion:
                    type: string
        '502':
          $ref: '#/components/responses/CoreFailure'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '403':
          $ref: '#/components/responses/Forbidden'
  /cores/{coreId}/actions/reconnect:
    post:
      tags: [ Cores ]
      operationId: reconnectCore
      summary: Drop the current connection and reconnect a Core
      security:
        - cookieAuth: [ ]
          csrfToken: [ ]
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/CoreId'
      responses:
        '202':
          description: Reconnect scheduled
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Core'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mcnp_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  headers:
    ETag:
      description: Quoted resource revision
      schema:
        type: string
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Opaque pagination cursor
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    CoreId:
      name: coreId
      in: path
      required: true
      schema:
        type: string
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthenticated:
      description: Authentication required or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource does not exist or is not visible
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource or state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RevisionMismatch:
      description: If-Match does not match current revision
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    CoreFailure:
      description: Core authentication or protocol failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Panel or Core is temporarily unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CoreStatus:
      type: string
      enum: [ ONLINE, DEGRADED, OFFLINE, INCOMPATIBLE, AUTH_FAILED, UNKNOWN ]
    Core:
      type: object
      required:
        - id
        - name
        - address
        - status
        - capabilities
        - secretConfigured
        - skipCertificateVerification
        - tags
        - revision
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
          description: IP/localhost host:port or a tls, mcnp, or https URL
        status:
          $ref: '#/components/schemas/CoreStatus'
        latencyMs:
          type: [ integer, 'null' ]
          minimum: 0
        lastSeenAt:
          type: [ string, 'null' ]
          format: date-time
        version:
          type: [ string, 'null' ]
        protocolVersion:
          type: [ string, 'null' ]
        capabilities:
          type: array
          items:
            type: string
        secretConfigured:
          type: boolean
        secretUpdatedAt:
          type: [ string, 'null' ]
          format: date-time
        skipCertificateVerification:
          type: boolean
        certificateVerified:
          type: [ boolean, 'null' ]
        tlsCertificateSha256:
          type: [ string, 'null' ]
          pattern: '^[0-9a-f]{64}$'
        tags:
          type: array
          items:
            type: string
        revision:
          type: integer
          minimum: 1
    CoreCreate:
      type: object
      required: [ name, address, secret ]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        address:
          type: string
          description: IP/localhost host:port or a tls, mcnp, or https URL
        secret:
          type: string
          minLength: 32
          writeOnly: true
        connectTimeoutSeconds:
          type: integer
          minimum: 1
          maximum: 60
          default: 10
        skipCertificateVerification:
          type: boolean
          default: false
        tags:
          type: array
          maxItems: 32
          items:
            type: string
    CoreUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        address:
          type: string
        secret:
          type: string
          minLength: 32
          writeOnly: true
        connectTimeoutSeconds:
          type: integer
          minimum: 1
          maximum: 60
        skipCertificateVerification:
          type: boolean
        tags:
          type: array
          maxItems: 32
          items:
            type: string
    CorePage:
      type: object
      required: [ items, nextCursor ]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Core'
        nextCursor:
          type: [ string, 'null' ]
...
    Error:
      type: object
      required: [ code, message, requestId, retryable ]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [ error ]
      properties:
        error:
          $ref: '#/components/schemas/Error'